+++ packages/core/src/action/index.ts
@@ -30,6 +30,7 @@ export interface IExecutionContext<
     : { machineContext?: TContext };
   runId?: string;
   mastra?: MastraPrimitives;
+  suspend: () => void;
 }
 export interface IAction<
   TId extends string,