+++ packages/core/src/workflows/types.ts
@@ -41,7 +41,10 @@ export type StepGraph = {
 
 export type RetryConfig = { attempts?: number; delay?: number };
 
-export type VariableReference<TStep extends StepVariableType<any, any, any, any>, TTriggerSchema extends z.ZodType<any>> =
+export type VariableReference<
+  TStep extends StepVariableType<any, any, any, any>,
+  TTriggerSchema extends z.ZodType<any>,
+> =
   TStep extends IAction<any, any, any, any>
     ? {
         step: TStep;
@@ -50,14 +53,17 @@ export type VariableReference<TStep extends StepVariableType<any, any, any, any>
     : TStep extends 'trigger'
       ? {
           step: 'trigger';
-          path: PathsToStringProps<ExtractSchemaType<TTriggerSchema>>;
+          path: PathsToStringProps<ExtractSchemaType<TTriggerSchema>> | '.' | '';
         }
       : {
           step: { id: string };
           path: string;
         };
 
-export interface BaseCondition<TStep extends StepVariableType<any, any, any, any>, TTriggerSchema extends z.ZodType<any>> {
+export interface BaseCondition<
+  TStep extends StepVariableType<any, any, any, any>,
+  TTriggerSchema extends z.ZodType<any>,
+> {
   ref: TStep extends IAction<any, any, any, any>
     ? {
         step: TStep;
@@ -66,7 +72,7 @@ export interface BaseCondition<TStep extends StepVariableType<any, any, any, any
     : TStep extends 'trigger'
       ? {
           step: 'trigger';
-          path: PathsToStringProps<ExtractSchemaType<TTriggerSchema>>;
+          path: PathsToStringProps<ExtractSchemaType<TTriggerSchema>> | '.' | '';
         }
       : {
           step: { id: string };
@@ -111,7 +117,9 @@ export interface StepConfig<
   TTriggerSchema extends z.ZodType<any>,
 > {
   snapshotOnTimeout?: boolean;
-  when?: Condition<CondStep, TTriggerSchema> | ((args: { context: WorkflowContext }) => Promise<boolean>);
+  when?:
+    | Condition<CondStep, TTriggerSchema>
+    | ((args: { context: WorkflowContext<TTriggerSchema> }) => Promise<boolean>);
   variables?: StepInputType<TStep, 'inputSchema'> extends never
     ? Record<string, VariableReference<VarStep, TTriggerSchema>>
     : {
@@ -139,9 +147,9 @@ type StepFailure = {
 export type StepResult<T> = StepSuccess<T> | StepFailure | StepSuspended | StepWaiting;
 
 // Update WorkflowContext
-export interface WorkflowContext<TTrigger = any> {
+export interface WorkflowContext<TTrigger extends z.ZodType<any> = any> {
   stepResults: Record<string, StepResult<any>>;
-  triggerData: TTrigger;
+  triggerData: z.infer<TTrigger>;
   attempts: Record<string, number>;
 }