+++ packages/core/src/workflows/workflow.ts
@@ -7,6 +7,7 @@ import { IAction, MastraPrimitives } from '../action';
 import { MastraBase } from '../base';
 import { LogLevel } from '../logger';
 
+import { WorkflowSnapshot } from './snapshot';
 import { Step } from './step';
 import {
   ActionContext,
@@ -35,6 +36,7 @@ export class Workflow<
   TTriggerSchema extends z.ZodType<any> = any,
 > extends MastraBase {
   name: string;
+  snapshot?: WorkflowSnapshot;
   triggerSchema?: TTriggerSchema;
   /** XState machine instance that orchestrates the workflow execution */
   #machine!: ReturnType<typeof this.initializeMachine>;
@@ -77,6 +79,11 @@ export class Workflow<
     this.triggerSchema = triggerSchema;
     this.#runId = crypto.randomUUID();
     this.#mastra = mastra;
+
+    if (this.#mastra?.engine) {
+      this.snapshot = new WorkflowSnapshot(this.#mastra.engine);
+    }
+
     if (mastra?.logger) {
       this.logger = mastra?.logger;
     }
@@ -234,9 +241,12 @@ export class Workflow<
   async execute({
     triggerData,
     loadSnapshot,
+    onStep,
   }: {
     triggerData?: z.infer<TTriggerSchema>;
     loadSnapshot?: { runId: string };
+    //TODO: TYPE
+    onStep?: (stepResults: any) => void;
   } = {}): Promise<{
     triggerData?: z.infer<TTriggerSchema>;
     results: Record<string, StepResult<any>>;
@@ -371,13 +381,16 @@ export class Workflow<
 
       const suspendedPaths: Set<string> = new Set();
 
-      console.log({ suspendedPaths }, '=======================');
-
       this.#actor.subscribe(state => {
         this.log(LogLevel.DEBUG, 'State update', {
           value: state.value,
           context: state.context,
           runId: this.#runId,
+        });
+
+        onStep?.({
+          step: state.value,
+          suspendedPaths,
         });
 
         this.#getSuspendedPaths({
@@ -408,6 +421,7 @@ export class Workflow<
 
         // Check if any steps failed
         const allStepsFailed = Object.values(state.context.stepResults).every(result => result.status === 'failed');
+
         const hasSuspended =
           Object.values(state.context.stepResults).some(result => result.status === 'suspended') ||
           suspendedPaths.size > 0;
@@ -639,6 +653,21 @@ export class Workflow<
                 { type: 'spawnSubscribers', params: { stepId: stepNode.step.id } },
               ],
             },
+            on: {
+              SUSPENDED: {
+                target: 'suspended',
+                actions: [
+                  assign({
+                    stepResults: ({ context }) => ({
+                      ...context.stepResults,
+                      [stepNode.step.id]: {
+                        status: 'suspended',
+                      },
+                    }),
+                  }),
+                ],
+              },
+            },
             onError: {
               target: 'failed',
               actions: [{ type: 'setStepError', params: { stepId: stepNode.step.id } }],
@@ -799,6 +828,10 @@ export class Workflow<
             machineContext: context,
             ...resolvedData,
           },
+          suspend: () => {
+            console.log('Calling SUSPEND');
+            this.#actor?.send({ type: 'SUSPENDED', stepId: stepNode.step.id });
+          },
           ...injectables,
         });
 
@@ -929,27 +962,29 @@ export class Workflow<
    * Persists the workflow state to the database
    */
   async #persistWorkflowSnapshot() {
-    const snapshot = this.#actor?.getPersistedSnapshot();
+    const snapshotFromActor = this.#actor?.getPersistedSnapshot();
+    if (!this.snapshot) {
+      this.log(LogLevel.ERROR, 'Snapshot cannot be persisted. Mastra engine is not initialized', {
+        runId: this.#runId,
+      });
+      return;
+    }
+
+    if (!snapshotFromActor) {
+      this.log(LogLevel.ERROR, 'Snapshot cannot be persisted. No snapshot received.', { runId: this.#runId });
+      return;
+    }
+
     this.log(LogLevel.INFO, 'Persisting workflow snapshot', {
-      snapshot,
+      snapshot: snapshotFromActor,
       runId: this.#runId,
     });
 
-    if (!this.#mastra) return;
-
-    const engine = this.#mastra.engine;
-
-    if (!engine || !snapshot) return;
-
-    await engine.syncRecords({
-      name: this.#entityName,
+    await this.snapshot.persist({
+      runId: this.#runId,
+      snapshot: snapshotFromActor,
+      entityName: this.#entityName,
       connectionId: this.#connectionId,
-      records: [
-        {
-          externalId: this.#runId,
-          data: { snapshot: JSON.stringify(snapshot) },
-        },
-      ],
     });
 
     this.log(LogLevel.INFO, 'Successfully persisted workflow snapshot', { runId: this.#runId });
@@ -957,11 +992,10 @@ export class Workflow<
   }
 
   async #loadWorkflowSnapshot(runId: string) {
-    if (!this.#mastra) return;
-
-    const engine = this.#mastra.engine;
-
-    if (!engine) return;
+    if (!this.snapshot) {
+      this.log(LogLevel.ERROR, 'Snapshot cannot be loaded. Mastra engine is not initialized', { runId });
+      return;
+    }
 
     this.log(LogLevel.INFO, 'Loading workflow snapshot', {
       runId,
@@ -969,19 +1003,18 @@ export class Workflow<
       connectionId: this.#connectionId,
     });
 
-    const state = await engine.getRecordsByEntityNameAndExternalId({
+    const snapshotData = await this.snapshot.load({
+      runId,
       entityName: this.#entityName,
       connectionId: this.#connectionId,
-      externalId: runId,
     });
 
     this.log(LogLevel.INFO, 'Retrieved workflow state from storage', {
-      state,
-      snapshot: state[0]?.data.snapshot,
+      snapshot: snapshotData,
       runId,
     });
 
-    return state[0]?.data.snapshot;
+    return snapshotData;
   }
 
   /**
@@ -1030,7 +1063,7 @@ export class Workflow<
       }
 
       // If path is empty or '.', return the entire source data
-      const value = variable.path === '' || variable.path === '.' ? sourceData[key] : get(sourceData, variable.path);
+      const value = variable.path === '' || variable.path === '.' ? sourceData : get(sourceData, variable.path);
 
       this.log(LogLevel.DEBUG, `Resolved variable ${key}`, {
         value,
@@ -1216,6 +1249,10 @@ export class Workflow<
 
     if (p.logger) {
       this.__setLogger(p.logger);
+    }
+
+    if (p?.engine) {
+      this.snapshot = new WorkflowSnapshot(p.engine);
     }
 
     this.#mastra = p;