+++ examples/bird-checker-with-express/CHANGELOG.md
@@ -1,5 +1,31 @@
 # bird-checker-with-express
 
+## 1.0.1-alpha.33
+
+### Patch Changes
+
+- Updated dependencies [eedb829]
+  - @mastra/core@0.1.27-alpha.52
+  - @mastra/engine@0.0.5-alpha.46
+
+## 1.0.1-alpha.32
+
+### Patch Changes
+
+- Updated dependencies [a7b016d]
+- Updated dependencies [da2e8d3]
+- Updated dependencies [538a136]
+  - @mastra/core@0.1.27-alpha.51
+  - @mastra/engine@0.0.5-alpha.45
+
+## 1.0.1-alpha.31
+
+### Patch Changes
+
+- Updated dependencies [401a4d9]
+  - @mastra/core@0.1.27-alpha.50
+  - @mastra/engine@0.0.5-alpha.44
+
 ## 1.0.1-alpha.30
 
 ### Patch Changes