+++ examples/workflow/CHANGELOG.md
@@ -1,5 +1,31 @@
 # workflow
 
+## 1.0.1-alpha.12
+
+### Patch Changes
+
+- Updated dependencies [eedb829]
+  - @mastra/core@0.1.27-alpha.52
+  - @mastra/memory@0.0.2-alpha.32
+
+## 1.0.1-alpha.11
+
+### Patch Changes
+
+- Updated dependencies [a7b016d]
+- Updated dependencies [da2e8d3]
+- Updated dependencies [538a136]
+  - @mastra/core@0.1.27-alpha.51
+  - @mastra/memory@0.0.2-alpha.31
+
+## 1.0.1-alpha.10
+
+### Patch Changes
+
+- Updated dependencies [401a4d9]
+  - @mastra/core@0.1.27-alpha.50
+  - @mastra/memory@0.0.2-alpha.30
+
 ## 1.0.1-alpha.9
 
 ### Patch Changes