+++ packages/engine/CHANGELOG.md
@@ -1,5 +1,28 @@
 # @mastra/engine
 
+## 0.0.5-alpha.46
+
+### Patch Changes
+
+- Updated dependencies [eedb829]
+  - @mastra/core@0.1.27-alpha.52
+
+## 0.0.5-alpha.45
+
+### Patch Changes
+
+- Updated dependencies [a7b016d]
+- Updated dependencies [da2e8d3]
+- Updated dependencies [538a136]
+  - @mastra/core@0.1.27-alpha.51
+
+## 0.0.5-alpha.44
+
+### Patch Changes
+
+- Updated dependencies [401a4d9]
+  - @mastra/core@0.1.27-alpha.50
+
 ## 0.0.5-alpha.43
 
 ### Patch Changes